+++ app/process/barriers/barriers.sql
@@ -1,0 +1,1 @@
+--TODO: cut out all bridges with bicycle and footway acess
